requirements.txt
python-telegram-bot==13.7
passport-eye==0.3.1
opencv-python-headless==4.5.5.64
pytesseract==0.3.9
Pillow==9.1.0
Flask==2.1.1